passssat/pyproject.toml
[tool.poetry]
name = "pass-at-app"
version = "0.1.0"
description = ""
authors = ["Michał Mastej <m.mastej1812@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

fastapi = "0.111.0"
jinja2 = "3.1.4"
uvicorn = "0.29.0"
python-jose = "3.4.0"
passlib = "1.7.4"
pandas = "2.2.3"
httpx = "^0.28.1"
itsdangerous = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
